Fix for review comments

+++ .github/scripts/clean-env-windows.ps1
@@ -45,7 +45,7 @@ $CleanEnvWindowsAllowList = @(
     "VCPKG_ROOT"
 )
 Get-ChildItem env: | ForEach-Object {
-    if ($CleanEnvWindowsAllowList -notcontains $_.Name -and $_.Name -notlike "ACTIONS_*" -and $_.Name -notlike "GIT_*" -and $_.Name -notlike "GITHUB_*" -and $_.Name -notlike "PLUGIN_*" -and $_.Name -notlike "RUNNER_*") {
+    if ($CleanEnvWindowsAllowList -notcontains $_.Name -and $_.Name -notlike "ACTIONS_*" -and $_.Name -notlike "CCACHE_*" -and $_.Name -notlike "GIT_*" -and $_.Name -notlike "GITHUB_*" -and $_.Name -notlike "PLUGIN_*" -and $_.Name -notlike "RUNNER_*") {
         Remove-Item $_.PSPath -Force
     }
 }
